requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
distro==1.9.0
filelock==3.15.4
Flask==3.0.3
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
fsspec==2024.6.1
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
llvmlite==0.43.0
m3u8==6.0.0
MarkupSafe==2.1.5
more-itertools==10.4.0
moviepy==1.0.3
mpmath==1.3.0
mutagen==1.47.0
networkx==3.3
numba==0.60.0
numpy==2.0.1
openai==1.42.0
openai-whisper==20231117
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
praw==7.7.1
prawcore==2.4.0
proglog==0.1.10
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PySocks==1.7.1
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
regex==2024.7.24
requests==2.32.3
selenium==4.23.1
setuptools==75.1.0
simple-websocket==1.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
sympy==1.13.2
tiktoken==0.7.0
torch==2.4.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.2.2
Wand==0.6.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.0
Werkzeug==3.0.4
wsproto==1.2.0
yt-dlp==2024.8.6